changed structure of folders, removed the responsability of connect do retry, updated readme and create dev.sh
