+++ .gitignore
@@ -1,2 +1,1 @@
-node/node_modules
-rabbitmq/
+node_modules
