--- node/index.js
@@ -1,1 +1,0 @@
-require("./interfaces/amqp/queues")
