+++ src/index.js
@@ -1,0 +1,3 @@
+require("./interfaces/amqp/queues");
+
+console.log("hello world")
