+++ src/interfaces/amqp/queues/shipments/index.js
